shared/requirements.txt
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
hypothesis>=6.0,<7.0
PyJWT>=2.8,<3.0
pyotp>=2.9,<3.0
